add version labeling, update content
