+++ versioned_docs/version-en/index.md
@@ -1,7 +1,6 @@
 ---
 id: about-mlapi
 title: About MLAPI
-slug: /
 ---
 
 The MLAPI is a high level networking library built for the Unity game engine to abstract networking. This allows developers to focus on the game rather than low level protocols and networking frameworks. All this is provided for free under the permissive MIT Licence by Unity and the MLAPI collaborators. The MLAPI is open source with no attached costs or limitations.
