+++ versioned_sidebars/version-(id)-sidebars.json
@@ -1,0 +1,151 @@
+{
+  "version-(id)/docs": [
+    {
+      "collapsed": true,
+      "type": "category",
+      "label": "Getting Started",
+      "items": [
+        {
+          "type": "doc",
+          "id": "version-(id)/about-mlapi"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/getting-started/installation"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/getting-started/library-initialization"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/getting-started/connection-approval"
+        }
+      ]
+    },
+    {
+      "collapsed": true,
+      "type": "category",
+      "label": "MLAPI Basics",
+      "items": [
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/messaging-system"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/object-spawning"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/object-behavior-relation"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/modularity"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/networkedvar"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/syncedvar"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/scene-management"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/object-ownership"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/object-visibility"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/mlapi-basics/logging"
+        }
+      ]
+    },
+    {
+      "collapsed": true,
+      "type": "category",
+      "label": "Core Components",
+      "items": [
+        {
+          "type": "doc",
+          "id": "version-(id)/core-components/core-components"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/core-components/networkingmanager"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/core-components/trackedobject"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/core-components/networkedobject"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/core-components/networkedbehavior"
+        }
+      ]
+    },
+    {
+      "collapsed": true,
+      "type": "category",
+      "label": "Advanced Topics",
+      "items": [
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/message-encryption"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/object-pooling"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/lag-compensation"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/bitwriter-bitreader-bitstream"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/custom-transports"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/networkprofiler-window"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/custom-serialization"
+        },
+        {
+          "type": "doc",
+          "id": "version-(id)/advanced-topics/boxing-systems"
+        }
+      ]
+    },
+    {
+      "collapsed": true,
+      "type": "category",
+      "label": "Troubleshooting",
+      "items": [
+        {
+          "type": "doc",
+          "id": "version-(id)/troubleshooting/troubleshooting"
+        }
+      ]
+    }
+  ]
+}
